functions/cost-estimation/requirements.txt
google-api-core==2.16.2
google-api-python-client==2.116.0
google-cloud-firestore==2.14.0
google-cloud-logging==3.9.0
functions-framework==3.5.0
pytest==8.0.1
mock-firestore==0.11.0
mock==5.1.0
flask==2.1.0